Foreign key changes for audit columns

+++ Duc.Splitt/Duc.Splitt.Data/DataAccess/Models/OtpRequest.cs
@@ -43,14 +43,14 @@ public partial class OtpRequest
     [Column(TypeName = "datetime")]
     public DateTime CreatedOn { get; set; }
 
-    public byte CreatedAt { get; set; }
+    public int CreatedAt { get; set; }
 
     public Guid? ModifiedBy { get; set; }
 
     [Column(TypeName = "datetime")]
     public DateTime? ModifiedOn { get; set; }
 
-    public byte? ModifiedAt { get; set; }
+    public int? ModifiedAt { get; set; }
 
     [InverseProperty("OtpRequest")]
     public virtual ICollection<CustomerRegistrationRequest> CustomerRegistrationRequest { get; set; } = new List<CustomerRegistrationRequest>();
